type fixed

+++ src/main/java/core/basesyntax/Ball.java
@@ -4,7 +4,7 @@ public class Ball {
     private String color;
     private byte number;
 
-    public Ball(String color, double number) {
+    public Ball(String color, byte number) {
         this.color = color;
         this.number = (byte) number;
     }
